requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
python-decouple==3.8
python-steam-api==1.2.2
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
yarl==1.9.4
